Rename ExampleFlail to ExampleAdvancedFlail
