+++ ExampleMod/Content/Items/Weapons/ExampleAdvancedFlail.png
