+++ ExampleMod/Content/Projectiles/ExampleAdvancedFlailProjectile.png
