+++ ExampleMod/Content/Projectiles/ExampleAdvancedFlailProjectileChain.png
