+++ ExampleMod/Content/Projectiles/ExampleAdvancedFlailProjectileChainExtra.png
